requirements.txt
fastapi
aiologger[aiofiles]
pip-chill
uvicorn
gunicorn
python-dotenv
requests
openai
faiss-cpu
PyJWT
cryptography
googlesearch-python
langchain-community
langchain_openai
langchain_core
aiohttp
markdownify